editUser on same url
