+++ tu-makerspace-card-reader/client/src/App.js
@@ -2,8 +2,7 @@ import './App.css';
 import MachineView from './MachineView.js';
 import AddUser from './AddUser.js';
 import { Routes, Route, Link } from 'react-router-dom';
-import EditUser1 from './EditUser1.js';
-import EditUser2 from './EditUser2.js';
+import EditUser from './EditUser.js';
 
 function Home(){
   return(
@@ -19,7 +18,7 @@ function Home(){
         <form action="/addUser">
           <button className = "Box">Add a User</button>
         </form>
-        <form action="/editUser1">
+        <form action="/editUser">
           <button className="Box">Edit User</button>
         </form>
       </nav>
@@ -49,8 +48,7 @@ function App() {
     {/* Maddie add a path for add user when u wanna start working on that and you now have a new page for that function! */}
       <Route path= "addUser" element = {<AddUser />} />
       <Route path="successfullyAddedUser" element = {<AddedUser />}/>
-      <Route path="editUser1" element = {<EditUser1 />}/>
-      <Route path="editUser2" element = {<EditUser2 />}/>
+      <Route path="editUser" element = {<EditUser />}/>
     </Routes>
     </div>
   );
